Crate opendal_layer_otelmetrics

Apache OpenDAL™ OpenTelemetry Metrics Layer

opendal-layer-otelmetrics records OpenDAL operation and HTTP metrics with OpenTelemetry.

Use through opendal

Applications should normally enable this layer through the opendal facade with the layers-otel-metrics feature:

cargo add opendal --features layers-otel-metrics

The layer is available as opendal::layers::OtelMetricsLayer and can be composed with an operator through opendal::Operator::layer.

Use with opendal-core

Add the split crates directly:

cargo add opendal-core opendal-layer-otelmetrics

Pass a configured layer to Operator::layer:

use opendal_core::Operator;
use opendal_layer_otelmetrics::OtelMetricsLayer;

fn add_layer(operator: Operator, layer: OtelMetricsLayer) -> Operator {
    operator.layer(layer)
}

Construct and configure layer as described in the API documentation before composing it with an operator.

Documentation

Rust API documentation

Apache OpenDAL production guide

Apache OpenDAL documentation

License

Licensed under the Apache License, Version 2.0.

Structs

OtelMetricsLayer

OtelMetricsLayer records OpenDAL operation and HTTP fetch metrics with [opentelemetry::metrics].

OtelMetricsLayerBuilder

OtelMetricsLayerBuilder is a config builder to build a OtelMetricsLayer.